Updated/added functions, flake8 compliance
Major updates include adding helper functions to help generalize the use of the module. The helper functions are countDueDates and countCells which resulted in a new implimentation of tabText. The script from the main file was moved to this file to consolidated the functions.

The code was updated to comply with flake8 recommendations with the exception of a few personal preferences.